+++ app/Entity/Shop/Catalog/Category/CategoryObserver.php
@@ -46,7 +46,9 @@ class CategoryObserver
      */
     public function deleted(Category $category)
     {
-        $this->removeFile($category->image);
+        if ($category->image) {
+            $this->removeFile($category->image);
+        }
     }
 
     /**
